refactor(overall): barrel-export OverallFactsSummary (review Minor, #546 consistency)
경량 순수 서버 컴포넌트라 @/widgets/overall barrel로 노출하고 page/test가 barrel 경유 import.
PR #546의 TechnicalFactsSummary와 동일 패턴 — 직접 lib import 회피.
