+++ src/widgets/overall/index.ts
@@ -1,6 +1,9 @@
 // overall widget barrel — public API for external consumers.
 
 export { OverallContent } from './OverallContent';
+// 경량 순수 서버 컴포넌트(types + JSX, client/heavy 의존 없음) — barrel로 노출한다.
+// app route의 종합 분석 FactLayer SSR fallback이 소비.
+export { OverallFactsSummary } from './OverallFactsSummary';
 export { OverallTriggerCta } from './OverallTriggerCta';
 export { ReanalyzeButton } from './ReanalyzeButton';
 export { DependencyProgress } from './DependencyProgress';
